+++ src/textual/layouts/vertical.py
@@ -20,6 +20,7 @@ class VerticalLayout(Layout):
     def arrange(
         self, parent: Widget, children: list[Widget], size: Size
     ) -> ArrangeResult:
+        parent.pre_layout(self)
         placements: list[WidgetPlacement] = []
         add_placement = placements.append
         viewport = parent.app.size
@@ -97,14 +98,17 @@ class VerticalLayout(Layout):
                 else NULL_OFFSET
             )
 
+            region = _Region(
+                box_margin.left,
+                y.__floor__(),
+                content_width.__floor__(),
+                next_y.__floor__() - y.__floor__(),
+            )
+            if styles.has_rule("position") and styles.position == "absolute":
+                region = region.reset_offset
             add_placement(
                 _WidgetPlacement(
-                    _Region(
-                        box_margin.left,
-                        y.__floor__(),
-                        content_width.__floor__(),
-                        next_y.__floor__() - y.__floor__(),
-                    ),
+                    region,
                     offset,
                     box_margin,
                     widget,